Remove commas when generating multiple Allow/Disallow rules
Fix a bug to remove unwanted commas from the output when having
multiple Allow and/or Disallow for the same agent.

Using the string += operator to add the Array.map output to a string
results in unwanted commas added to the string.

Added unit tests too to cover this edge case